Added new number drawing algorithm.
