+++ color_clustering.py
@@ -1,3 +1,4 @@
+from collections import deque
 import cv2 as cv
 import numpy as np
 import matplotlib.pyplot as plt
@@ -36,7 +37,11 @@ def k_means_clustering(k, img):
         color_pallete.append(centroids[i])
     # display_image(clustered_img, "Clustered image")
     print("Image successfully generated")
-    return clustered_img, clusters, color_pallete
+
+    print("Generating batches...")
+    batches, center_of_masses = generate_batches(clustered_img, clusters, color_pallete)
+    
+    return clustered_img, clusters, color_pallete, batches, center_of_masses
 
     
 #Assigns all pixels to the cluster center they are closest to
@@ -94,3 +99,48 @@ def pick_centroids(k, img):
     centroids = np.array(centroids, dtype=np.uint8)
     return centroids, clusters #we now return k cluster centers that are spawned on random points in our RGB space
     
+def generate_batches(img, clusters, color_pallete):
+    visited = np.zeros((img.shape[0], img.shape[1], 1), dtype=bool)
+    directions = [(-1, 0), (1, 0), (0, -1), (0, 1)] #used to check up down left right easier
+    batches_within_cluster = [[] for _ in range(len(clusters))]
+    center_of_mass = [[] for _ in range(len(clusters))]
+    h_bound, w_bound, _ = img.shape
+    
+    for cluster_index, cluster in enumerate(clusters):
+        #remember that cluster is an array that holds x,y values to pixels of that color
+        cluster_color = color_pallete[cluster_index] #get the color value
+        for coord in cluster:
+            x, y = coord
+            if(visited[x][y] == True): #Dont start queue if alr searched this pixel
+                continue
+            queue = deque()
+            queue.append((x, y))
+            
+            #checks each cluster within the cluster
+            batch = [] #holds the coords for all pixels within the specific batch in the cluster
+            batch.append((x, y))
+            x_center = y_center = 0
+            num_pixels = 0
+            while queue:
+                x,y = queue.pop()
+                visited[x][y] = True
+                for dx, dy in directions:
+                    nx, ny = x + dx, y + dy
+                    if 0 <= nx < h_bound and 0 <= ny < w_bound: #bounds checking
+                        #if the adj pixel we are looking at is the same value as current & havent visted
+                        # print("IMG",img[x][y])
+                        # print("CLUSTER COLOR", cluster_color)
+                        if((img[nx][ny] == cluster_color).all() and visited[nx][ny] == False):
+                            num_pixels += 1
+                            queue.append((nx, ny))
+                            batch.append((nx, ny))
+                            x_center += nx
+                            y_center += ny
+            if(num_pixels > 100): #remove tiny little bits from having numbers
+                x_center_final = x_center // num_pixels
+                y_center_final = y_center // num_pixels
+                center_of_mass[cluster_index].append((int(x_center_final), int(y_center_final)))
+                if(x_center_final < 0):
+                    print(x_center, y_center, num_pixels)
+            batches_within_cluster[cluster_index].append(batch) #for cluster idx append all batches found within
+    return batches_within_cluster, center_of_mass
